pyproject.toml
[project]
name = "advent-of-code"
version = "0.1.0"
description = "Advent of Code solutions"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

